+++ NEWS.md
@@ -1,6 +1,10 @@
 # rasterpic (development version)
 
--   Add support for `SpatVector`, `SpatExtent`, `sf::st_bbox()`.
+-   Add support for
+
+    -   **terra**: `SpatVector`, `SpatExtent`,
+
+    -   **sf**: `sfg`, `sf::st_bbox()`.
 
 # rasterpic 0.1.0